SSC_Site/requirements.txt
Mezzanine==4.0.1
django-auth-ldap==1.2.7
django-modeltranslation==0.10.1
psycopg2==2.6.1
python-memcached==1.57
gunicorn==19.3.0
jdatetime==1.7.1
redis==2.10.5
django-celery-email==1.1.4
-e git://github.com/khodabande/Shortener.git#egg=shortener
zeep==0.14.0
Tempita==0.5.2